backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
redis[asyncio]==5.0.4
boto3==1.34.100
botocore==1.34.100
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
pgvector==0.2.5
sentence-transformers==3.0.0
python-dotenv==1.0.1
